Create separate module for icmp beacon
